Update tooltip.js
When I used a definition in the tooltip which didn't fit the width, the tooltip changed to wrapping the text, which is great, but the text ran outside the side of the browser window. Anyhow, this change makes certain that the tooltip will fit inside the browser window from top-to-bottom AND left-to-right. If it then doesn't fit the original intention, then the position is changed. The biggest challenge was making certain the tooltip's triangle stayed center even when the tooltip itself might no longer be centered. This was tested on Chrome, Safari, Firefox, Edge new/old, and iOS mobile.